refactor(render): demote pbrMaterial to test fixture; remove from public API
The previously-shipped pbrMaterial was not actual PBR (Blinn-Phong + Lambertian
+ 3% ambient floor, single hard-coded directional light). It is renamed to
litMaterial and relocated to tests/fixtures/lit-material/ as an internal demo
of how to build a custom material on top of the material() factory.

The library's public material surface is now generic only:
- material() factory (full WGSL control)
- Material interface (output type)
- RapidRenderer (generic over Material)

Users build their own PBR (or any shading model) in their own projects.

The lit-cube acceptance snapshot test still exists as an integration smoke test;
it imports from the new fixture path and produces byte-identical output
(sha256 unchanged).

Removed from public API:
- pbrMaterial factory
- PbrMaterial / PbrUniformValues / PbrMaterialSpec types
- DirectionalLight interface
- PBR shader source
- All "PBR" naming from user-facing identifiers

Bundle: @vgpu/render (.) production drops by 1630 B gz (PBR shader source +
factory no longer ship to consumers).
