+++ packages/render/src/inspect/inspect-material.docs.md
@@ -16,7 +16,7 @@ material.writeUniforms(uniforms.gpu, 0, {
 ```
 
 `writeUniforms` writes the shared inspect matrices into the buffer. This keeps
-inspect materials independent from the PBR material path used by `RapidRenderer`.
+inspect materials independent from application materials used by `RapidRenderer`.
 
 `InspectMaterialUniformParams` includes `viewProjectionMatrix` and `modelMatrix`.
 Future inspect materials should extend this interface explicitly if they need
