+++ packages/render/tests/fixtures/lit-material/index.ts
@@ -1,0 +1,2 @@
+export { litMaterial } from "./lit-material.ts";
+export type { DirectionalLight, LitMaterial, LitMaterialSpec, LitUniformValues } from "./lit-material.ts";
