+++ scripts/generate-lit-cube-snapshot.ts
@@ -4,7 +4,8 @@ import { dirname, join } from "node:path";
 import { PNG } from "pngjs";
 import { createNodeAdapter } from "@vgpu/adapter-node";
 import { App } from "@vgpu/core";
-import { degToRad, Mesh, pbrMaterial, perspectiveCamera, RapidRenderer, srgb, type Mat4, type Vec3 } from "@vgpu/render";
+import { degToRad, Mesh, perspectiveCamera, RapidRenderer, srgb, type Mat4, type Vec3 } from "@vgpu/render";
+import { litMaterial } from "../packages/render/tests/fixtures/lit-material/index.ts";
 
 const WIDTH = 256;
 const HEIGHT = 256;
@@ -22,7 +23,7 @@ async function main(): Promise<void> {
   const depth = device.createTexture({ size: [WIDTH, HEIGHT], format: "depth24plus", usage: ["render_attachment"] });
 
   const mesh = Mesh.box({ device, size: 1 });
-  const material = pbrMaterial({ device, baseColor: srgb(0xcc8844), targetFormat: FORMAT });
+  const material = litMaterial({ device, baseColor: srgb(0xcc8844), targetFormat: FORMAT });
   const camera = perspectiveCamera({
     fovYRadians: degToRad(45),
     aspect: 1,
